requirements.txt
algoliasearch==3.0.0
algoliasearch-django==3.0.0
asgiref==3.6.0
attrs==23.1.0
bcrypt==4.0.1
blessed==1.20.0
boto3==1.26.153
botocore==1.29.153
cached-property==1.5.2
cachetools==5.3.0
cement==2.8.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==4.2
django-allauth==0.54.0
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==23.2
django-sslserver==0.22
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==3.2.0
MarkupSafe==2.1.2
oauthlib==3.2.2
paramiko==3.2.0
pathspec==0.10.1
Pillow==9.5.0
protobuf==4.22.3
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.3
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.4
termcolor==1.1.0
texttable==1.6.7
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.1.9
websocket-client==0.59.0
Werkzeug==2.3.0
xlrd==2.0.1
